requirements.txt
flask
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
peft
jinja2
python-dotenv
torch
pdfplumber
numpy
gunicorn
pymongo
scikit-learn
nltk
